Add files via upload
Battleship final for computer science intro class. Right now, creating the hit or miss function which will take a user entered coordinate and check it against the ship array to see if it's a hit or not. Right now, I'm working on converting the letter coordinate to a number coordinate from the user entry, and the converted letter coordinate and the user entered numeric coordinate to check against the correct location in the hit array.