Create placeShipsglitch
added several functions: getNumberCoordinate, getLetterCoordinate, hitCheckChar, hitCheckNum. 
Small edits to hitArray changes. Ships now go in as S, hits are X and misses are M
Currently having trouble placing more than one ship with placeShips. This is because either the same number is generated when calling getRandom or because it doesn't generate a new random number. Regardless of which it is, I'm at a loss as to how to fix it. 
Problem lines in question are  132 - 182.